Assesment_Project/requirements.txt
streamlit>=1.10.0
pandas>=1.3.0
google-auth>=2.6.0
gspread>=5.1.1
python-dotenv>=0.19.0
groq>=0.3.0
tenacity>=8.2.2